Rename census_queries.py to databricks_census_queries.py
Clarifies that these SQL queries are specific to Databricks SQL.
